+++ ddtrace/contrib/internal/logging/constants.py
@@ -1,0 +1,7 @@
+RECORD_ATTR_TRACE_ID = "dd.trace_id"
+RECORD_ATTR_SPAN_ID = "dd.span_id"
+RECORD_ATTR_ENV = "dd.env"
+RECORD_ATTR_VERSION = "dd.version"
+RECORD_ATTR_SERVICE = "dd.service"
+RECORD_ATTR_VALUE_ZERO = "0"
+RECORD_ATTR_VALUE_EMPTY = ""
