+++ ddtrace/contrib/logging/__init__.py
@@ -68,8 +68,12 @@ required_modules = ["logging"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.logging.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.logging.patch import get_version
+        from ..internal.logging.patch import patch
+        from ..internal.logging.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
