+++ ddtrace/contrib/loguru/__init__.py
@@ -70,8 +70,12 @@ required_modules = ["loguru"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.loguru.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.loguru.patch import get_version
+        from ..internal.loguru.patch import patch
+        from ..internal.loguru.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
