+++ releasenotes/notes/move-integrations-to-internal-langchain-2cd7ff9a1362182d.yaml
@@ -1,0 +1,8 @@
+---
+deprecations:
+  - |
+    langchain: Deprecates all modules in the ``ddtrace.contrib.langchain`` package. Use attributes exposed in ``ddtrace.contrib.langchain.__all__`` instead.
+  - |
+    logbook: Deprecates the ``ddtrace.contrib.logbook`` module. Use ``ddtrace.contrib.langchain.logbook`` instead.
+  - |
+    loguru: Deprecates the ``ddtrace.contrib.loguru`` module. Use ``ddtrace.contrib.langchain.loguru`` instead.
